Addition of StreamJSON class. At that point the exportToJSON method is tested and is working, but while testing the importFromJSON method there is an InvocationTargetException, probably related to the Deserialization of the JSON stream.
In order to implement the logic behind JSON parsing, tried to implement it using the json-simple toolkit but that resulted in many exceptions while testing it. Then, while searching a bit more on that task, found a workaround solution (based on the following article: https://www.mkyong.com/java/jackson-how-to-parse-json/) which uses the Jackson API instead.